app/requirements.txt
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
asyncio==3.4.3
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
DoubleMetaphone==1.1
duckdb==0.10.1
entrypoints==0.4
executing==2.0.1
fastjsonschema==2.19.1
fqdn==1.5.1
fsspec==2023.12.2
future==0.18.3
geocoder==1.38.1
h11==0.14.0
htmltools==0.5.1
humanize==4.9.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipysigma==0.24.2
ipython==8.18.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter_client==7.4.9
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
linkify-it-py==2.0.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
msgspec==0.18.6
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.26.3
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pins==0.8.4
platformdirs==4.2.0
probableparsing==0.0.1
probablepeople==0.5.4
prometheus-client==0.19.0
prompt-toolkit==3.0.36
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-crfsuite==0.9.10
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.4
PyYAML==6.0.1
pyzmq==24.0.1
questionary==2.0.1
ratelim==0.1.6
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsconnect==0.1.3
rsconnect-jupyter==1.8.0
rsconnect_python==1.22.0
semver==2.13.0
Send2Trash==1.8.2
shiny==0.7.0
shinywidgets==0.3.0
six==1.16.0
sniffio==1.3.0
sodapy==2.2.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.34.0
terminado==0.18.0
tinycss2==1.2.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uc-micro-py==1.0.2
uri-template==1.3.0
urllib3==2.2.0
us==3.1.1 
usaddress==0.5.10
usaddress-scourgify==0.6.0
uszipcode==1.0.1
uvicorn==0.27.0.post1
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
widgetsnbextension==4.0.9
xxhash==3.4.1
yaml-config==0.1.5
zipp==3.17.0
